Polish test methodology
